requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
distro==1.9.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.2
openai==1.100.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.2
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.4.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tk==0.1.0
tqdm==4.67.1
ttkbootstrap==1.14.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
